+++ client/src/MobicoopBundle/Resources/assets/js/components/user/profile/Profile.vue
@@ -100,6 +100,19 @@
               </v-tab-item>              
             </v-tabs>
           </v-tab-item>
+          <v-tab
+            v-if="showReviews"
+            class="text-left justify-start ml-2 mr-5 text-h6"
+            href="#reviews"
+          >
+            {{ $t("tabs.reviews") }}
+          </v-tab>
+          <v-tab-item
+            v-if="showReviews"
+            value="reviews"
+          >
+            <ReviewDashboard />
+          </v-tab-item>          
           <div>
             <ProfileSummary
               :user-id="user.id"
@@ -150,6 +163,7 @@ import CarpoolSettings from "@components/user/profile/CarpoolSettings";
 import BankAccount from "@components/user/profile/payment/BankAccount";
 import ProfileSummary from "@components/user/profile/ProfileSummary";
 import PublicProfile from "@components/user/profile/PublicProfile";
+import ReviewDashboard from "@components/user/profile/review/ReviewDashboard";
 
 import {messages_en, messages_fr} from "@translations/components/user/profile/Profile/";
 
@@ -168,7 +182,8 @@ export default {
     Carpools,
     BankAccount,
     ProfileSummary,
-    PublicProfile
+    PublicProfile,
+    ReviewDashboard
   },
   props: {
     user: {
@@ -226,7 +241,11 @@ export default {
     validationDocsAuthorizedExtensions: {
       type: String,
       default: null
-    }
+    },
+    showReviews: {
+      type: Boolean,
+      default: false
+    }    
   },
   data(){
     return{