+++ client/src/MobicoopBundle/User/Controller/UserController.php
@@ -393,6 +393,9 @@ class UserController extends AbstractController
         if ($tabDefault == 'mon-profil') {
             $tabDefault = 'myProfile';
         }
+        if ($tabDefault == 'avis') {
+            $tabDefault = 'reviews';
+        }
 
         return $this->render('@Mobicoop/user/updateProfile.html.twig', [
             'error' => $error,
@@ -401,7 +404,8 @@ class UserController extends AbstractController
             'ads' => $userManager->getAds(),
             'acceptedCarpools' => $userManager->getAds(true),
             'bankCoordinates' => $this->paymentElectronicActive,
-            'validationDocsAuthorizedExtensions' => $this->validationDocsAuthorizedExtensions
+            'validationDocsAuthorizedExtensions' => $this->validationDocsAuthorizedExtensions,
+            'showReviews' => $user->isUserReviewsActive()
         ]);
     }
 
@@ -1261,5 +1265,19 @@ class UserController extends AbstractController
             }
         }
         return new JsonResponse(["success"=>false]);
+    }
+
+    /**
+     * Get the Review Dashboard of a User
+     *
+     * @param Request $request
+     * @return void
+     */
+    public function userReviewDashboard(Request $request, ReviewManager $reviewManager)
+    {
+        if ($request->isMethod('POST')) {
+            return new JsonResponse($reviewManager->reviewDashboard());
+        }
+        return new JsonResponse();
     }
 }